requirements.txt
opencv-python>=4.1.1
pillow>=10.3.0
PyYAML>=5.3.1
torch>=1.6.0  
torchvision>=0.7.0
tqdm>=4.66.3
numpy
matplotlib
